+++ node/middlewares/routes/errorHandler.ts
@@ -6,25 +6,19 @@ export async function errorHandler(ctx: Context, next: () => Promise<void>) {
   try {
     await next()
   } catch (error) {
-    // If the Glovo order was received, add the order id to the log.
-    if (ctx.state.glovoOrder) {
-      const { order_id: orderId } = ctx.state.glovoOrder
+    // eslint-disable-next-line no-console
+    console.log({ error }, error.error)
 
-      logger.error({
-        orderId,
-        message: error.message,
-        error: error.resonse,
-      })
-    } else {
-      logger.error({
-        message: error.message,
-        status: error.status,
-        error: error.response,
-      })
-    }
+    logger.error({
+      glovoOrder: ctx.state?.glovoOrder ?? 'No Glovo order information',
+      message: error.message,
+      reason: error.reason,
+      metric: error.metric,
+      data: error.data,
+      error: error.error,
+    })
 
-    ctx.status = error.status || 500
-    ctx.body = error.message
+    ctx.body = error.reason
     ctx.app.emit('error', error, ctx)
   }
 }