+++ node/middlewares/routes/getGlovoMenuByStore.ts
@@ -1,4 +1,4 @@
-import { CustomError, generateStoreMenuRecord } from '../../utils'
+import { ServiceError, generateStoreMenuRecord } from '../../utils'
 
 export async function getGlovoMenuByStore(ctx: Context) {
   const {
@@ -13,11 +13,13 @@ export async function getGlovoMenuByStore(ctx: Context) {
 
     generateStoreMenuRecord(ctx, affiliateId)
   } catch (error) {
-    throw new CustomError({
-      message: `There was a problem getting the Glovo menu for the specified store`,
-      status: 500,
-      payload: error,
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason:
+        error.reason ??
+        `There was a problem getting the Glovo menu for the specified store`,
+      metric: 'menu',
+      error: error.response?.data,
     })
   }
 }