+++ node/middlewares/routes/storeMenuUpdate.ts
@@ -1,4 +1,4 @@
-import { CustomError } from '../../utils'
+import { ServiceError } from '../../utils'
 
 export async function storeMenuUpdates(ctx: Context) {
   const {
@@ -12,11 +12,13 @@ export async function storeMenuUpdates(ctx: Context) {
 
     ctx.body = storeMenuUpdatesRecord ?? 'Store menu updates record not found'
   } catch (error) {
-    throw new CustomError({
-      message: `There was a problem getting the Glovo menu updates for the specified store`,
-      status: 500,
-      payload: error,
-      error,
+    throw new ServiceError({
+      message: error.message,
+      reason:
+        error.reason ??
+        `There was a problem getting the Glovo menu updates for the specified store`,
+      metric: 'menu',
+      error: error.response?.data,
     })
   }
 }