+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Support for order changes.
 - Additional customer's information on VTEX orders.
+- Improvements for error logs.
 
 ## [3.1.0] - 2022-12-15