+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.2.4] - 2023-02-28
+
 ### Added
 
 - Support for order changes.
 - Additional customer's information on VTEX orders.
+- Improvements for error logs.
 
 ## [3.1.0] - 2022-12-15